// source/frontend/src/models/api/generated/ApiGen_Concepts_ManagementActivityInvolvedParty.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { ApiGen_Base_BaseAudit } from './ApiGen_Base_BaseAudit';
import { ApiGen_Concepts_ManagementActivity } from './ApiGen_Concepts_ManagementActivity';
import { ApiGen_Concepts_Organization } from './ApiGen_Concepts_Organization';
import { ApiGen_Concepts_Person } from './ApiGen_Concepts_Person';

// LINK: @backend/apimodels/Models/Concepts/Property/ManagementActivityInvolvedPartyModel.cs
export interface ApiGen_Concepts_ManagementActivityInvolvedParty extends ApiGen_Base_BaseAudit {
  id: number;
  organizationId: number | null;
  organization: ApiGen_Concepts_Organization | null;
  personId: number | null;
  person: ApiGen_Concepts_Person | null;
  managementActivityId: number;
  managementActivity: ApiGen_Concepts_ManagementActivity | null;
}